test(sandbox): consolidate start readiness coverage
